Refactor: harmonize gpio and ioc to new file name convention
