+++ example/firmware/config/RTE_Components.h
@@ -1,2 +1,3 @@
 
 #define RTE_Drivers_SPI // turn on the SPI driver
+#include "spi_pcbDetails.h"
