+++ example/firmware/config/spi_config.c
@@ -5,8 +5,9 @@
  * - Communicates this information to the spi driver.
  */
 #include <fmt_spi.h>
-#include <RTE_DeviceConfig.h> // PCB interconnect information. 
-#include <Device_IOC.h> // Target MCU interconnect information.
+#include "spi_pcbDetails.h" // replaces RTEDevice.h for ARM SPI driver
+#include "ioc_pcbDetails.h"
+#include <ioc_mcuDetails.h> // IOC_x_to_ISR_y mcu-supported routing options
 #include "ISR_Config.h"
 
 extern ARM_DRIVER_SPI Driver_SPI4;
@@ -14,9 +15,9 @@ extern ARM_DRIVER_SPI Driver_SPI4;
 spiCfg_t spiConfig = {
   .spiModuleId = 4,
   .spiModule = &Driver_SPI4,
-  .msgWaitingIocId = IOC_14_RTE_INDEX, // IOC_14: P1_0 in 4800_F144/Device_IOC.h
+  .msgWaitingIocId = IOC_14_RTE_INDEX, // IOC_14: P1_0 on XMC4700
   .msgWaitingIocOut = IOC_14_to_ISR_3,
-  .clearToSendIocId = IOC_4_RTE_INDEX, // IOC_4:  P0_4 in 4800_F144/Device_IOC.h
+  .clearToSendIocId = IOC_4_RTE_INDEX, // IOC_4:  P0_4 on XMC4700
   .clearToSendIocOut = IOC_4_to_ISR_2,
   .baudHz = 1000000,
   .busMode = BUS_MODE_MAIN,
