+++ firmware/fmt_gpio.h
@@ -4,15 +4,15 @@
  * access.
  * 
  * Mapping between pinId used here and target MCU ports/pins is done in 
- * <projectSrc>/config/gpio_pcb.h  (because it's pcb-determined information)
+ * <projectSrc>/config/gpio_pcbDetails.h  (because it's pcb-determined information)
  * 
  * Recommended use:
  * Init functions should be called from the modules that use these pins.
  * Configuring a new pin for GPIO use requires the following:
- * 1. In project/config/gpio_pcb.h, follow the instructions to add a new GPIO. 
+ * 1. In project/config/gpio_pcbDetails.h, follow the instructions to add a new GPIO. 
  * 2. Verify this new gpio functions on the intended hardware.
  * 3. Add #define for the macro added to AVAILABLE_GPIOs in step 1. to
- *    firment/firmware/port/host/mcu_gpio.h - This signifies passing step 2.
+ *    firment/firmware/port/host/gpio_mcuDetails.h - This signifies passing step 2.
  * 
  * Testing: 
  * The gpioTest.c will complain if step 3 above is not done.  It's a reminder
