+++ firmware/port/host/gpio_spy.c
