+++ todo.md
@@ -1,4 +1,11 @@
 # Doing 
+1. Use descriptive names in ioc_pcbDetails.h
+2. Excise "RTE" from ioc_pcbDetails.h
+3. Can I ditch the AVAILABLE_<FEAT>s tables? 
+  - maybe it allows for run-time selection of pins (for HW_ID changes)?
+4. Can I make a more performant interface for GPIO / IOC?
+  - Would it actually get better test-coverage?
+5. Move host/<feat>_mcuDetails.h files to project test dir.
 
 # Project structuring
 12. break fmt_rx -> project_comms dependency.
